Rename IcalFilterProxy.filters to .calendars
Returning a hash that maps calendar names to Calendar objects,
'calendars' is a more accurate description of the method and makes the
code more readable in various places.

Also rewrote the method slightly to use `transform_values` instead of
`each_with_object`, which is a more concise description of what the
method does.